requirements.txt
rdflib>=7.1.1
pyyaml
